+++ docker/entrypoint.sh
@@ -26,37 +26,6 @@ EOF
 
 fi
 
-if [ -f "serverConfig/server.js" ]
-then
-    echo "Config already exists"
-else
-    echo "" > serverConfig/server.js
-    tee serverConfig/server.js <<EOF >/dev/null
-module.exports = {
-  port: 4000,
-  registration: true,
-  loginRequired: true
-};
-EOF
-fi
-
-if [ -f "serverConfig/timeout.js" ]
-then
-    echo "Config already exists"
-else
-    echo "" > serverConfig/timeout.js
-    tee serverConfig/timeout.js <<EOF >/dev/null
-module.exports = {
-    webSocketRetry: 5000,
-    apiTimeout: 1000,
-    apiRetryCutoff: 10000,
-    apiRetry: 300000,
-};
-EOF
-fi
-
-
-
 if [ -d "logs" ]
 then
     mkdir -p logs