+++ routes/SSE.js
@@ -1,7 +1,6 @@
 const express = require("express");
 const router = express.Router();
 const { ensureAuthenticated } = require("../config/auth");
-const yj = require("yieldable-json");
 const ClientSettings = require("../models/ClientSettings.js");
 const {parse, stringify} = require('flatted/cjs');
 const ServerSettings = require("../models/ServerSettings.js");
@@ -14,7 +13,6 @@ const farmStatistics = require("../runners/statisticsCollection.js");
 const FarmStatistics = farmStatistics.StatisticsCollection;
 const runner = require("../runners/state.js");
 const Runner = runner.Runner;
-const Roll = require("../models/Filament.js");
 
 let clientId = 0;
 let clients = {}; // <- Keep a map of attached clients